Add builder for Flatcar
On Flatcar, a kernel version alone is not enough to determine the kernel config
and which GCC version was used. But a Flatcar release version number can will
give us this information, including the kernel version. So this build works the
following way: the kernel release field has to contain a flatcar version number
followed by `-flatcar`, such as 3185.0.0-flatcar. Based on this, the correct
"channel" will be found (there is 1-1 mapping between flatcar version and
channel), and we will fetch kernel version, GCC version and kernel config.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>