build/builder.Dockerfile: install dwarves/pahole into builder image
Flatcar kernels are built with BTF type info, which requires pahole when
configuring the kernel. If pahole is missing, then running `make
modules_prepare` removes entries from the kernel config and results in built
modules failing to probe with the error:

  module: x86/modules: Skipping invalid relocation target, existing value is nonzero for type 1, loc 0000000090295d6a, val ffffffffc0b5ae8d

Buster provides pahole 1.12, but CONFIG_PAHOLE_HAS_SPLIT_BTF requires 1.19
which, luckily, is available in the buster-packports repo. Add that package to
the container image.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>